update: album song ui
